pyproject.toml
[tool.poetry]
name = "meta-assistant"
version = "0.1.0"
description = ""
authors = ["Nicola Massarenti <nicola.massarenti@gmail.com>"]
readme = "README.md"
packages = [{include = "meta_assistant"}]

[tool.poetry.dependencies]
python = "^3.10"
typer = "^0.7.0"
loguru = "^0.6.0"
grpcio = "^1.51.1"
grpcio-tools = "^1.51.1"
numpy = "^1.24.1"
soundfile = "^0.11.0"
pydantic = "^1.10.4"
google-cloud-speech = "^2.17.1"
pyaudio = "^0.2.13"
six = "^1.16.0"
openai = "^0.26.4"
google-cloud-texttospeech = "^2.14.1"
webrtcvad = "^2.0.10"
sounddevice = "^0.4.5"
pydub = "^0.25.1"


[tool.poetry.group.dev.dependencies]
black = "^22.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
